inbounds on code blocks
Added inbounds to the first few functions, and moved the inbounds to the code blocks on others. Note that `@inbounds` propagates through for the whole code block so you can just put it on the `for` (but it doesn't propagate into functions!). Moving it like this also turned off inbounds checking on a few `w[i]` calls. Altogether, this likely won't affect performance much at all, but it makes it all more uniform.

Tests weren't run (do you have unit tests?)